Move `FromDyn` closer to `is_dyn_thread_safe`
